Add: created and tested PolygonCalculator.setPerimeter, .setCorrectrionsHorAngle

+++ geocalculator_service/src/main/java/ru/taheoport/geocalculator_service/model/Residuals.java
@@ -22,7 +22,7 @@ public class Residuals {
     /**
      * Actual angle residuals in seconds
      */
-    private double angle;
+    private long angle;
 
     /**
      * X-axis linear residuals in millimeters
